refactor covenant program
